kivy-protonox-version/deb_dist/protonox-kivy-3.0.0~dev8/examples/kivymd_demo/requirements.txt
kivymd>=1.1.1
kivy>=2.2.0
Pillow>=9.0.0
